% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compactify_edges}
\alias{compactify_edges}
\title{Construct indices matrices}
\usage{
compactify_edges(w, n, method = "ama")
}
\arguments{
\item{w}{weights vector}

\item{n}{number of points to cluster}

\item{method}{'ama' or 'admm'}
}
\description{
\code{compactify_edges} constructs M1, M2, and ix index matrices. Note that storage conventions are different for ama and admm.
}
\examples{
p <- 10
n <- 20
X <- matrix(rnorm(p*n),p,n)
w <- kernel_weights(X,0)
w <- knn_weights(w,3,n)
method <- 'admm'
edge_info <- compactify_edges(w,n,method=method)
method <- 'ama'
edge_info <- compactify_edges(w,n,method=method)
}

